Add Training Plan/Protocols test case(s) with full session reproduce (issue-409) (#419)
* Add Training Plan/Protocols test case(s) with full session reproduce (issue-409)

+ move behavior mock system to top_fixture for reuse
+ fix: inference model: do not close multiproc queue many times
+ fix: also disconnect Device in DeviceConnection
+ enh InferenceProtocol: add many missed function definitions
+ enh few logs

* Allow give extra args to GH pytest

* temporary

* Update checkout & setup-python version + put FORCE_COLOR as env.

* temporary

* Autouse diamond_config_path

* allow color output for pip install

* Use real diamond triangle config

* use calib dir

* auto-cancel old workflow if new commit

* good diamond_config_path

* pip install color

* join must only be done if all items have been task_done

* redo-all

* Fix test

* enh DeviceConnection + hardware model disconnect

* test

* file event plugin: fix: create dir if not exist

* Allow pass system machine to app model -> behavior model

* test training plan

* Better. do not leave `e` in class namespace behind.

* Allow patch InferenceModel private methods.

NB:

do not use double leading underscore style private methods as prohibited by Events

* main window: allow fully simulate intersession analysis

* put detection complete after detection_result_ready

* make handle detection result sync

* clean-test

* Enh/Fix: apply shift-xyz in inference coordinate system

* enh: show/hide dev analysis controls when enabled/not enabled

* Enh detection/analysis-results UI: prevent toolbar height change when toggled

* ShiftXYZHandler: expose handle_new_shift_xyz_func

* Fix: allow many AppModel.on_close

* use EventManager.try_close_default

* Allow float shift x/y/z

* Enh training plan test + fixtures

* Enh: allow coverage in subprocess

* better name

* missed-remove

* Smaller range for internal/dev shift-xyz result spinboxes

* enh training_plan test: make first assertion

* continue

* Use local cur_state already available, instead of rechecking property on self

* add debug

* Add log of UI HW controls/commands

* Enh/Fix: allow force CAN emulation interface

for test notably.

* update auto-trainer-training to 0.9.5 and auto-trainer-api to 0.9.3

* update training requirement 0.9.6

* Listen to training phase property changed

* update training to 0.9.7 + listen on plan progress_updated + better phase management

* fix: set before

* Fix: only show buttons in manual-and-protocol

* Again do plan deepcopy when using from get_training_plan_by_id

* Fix: reuse attache plan for switch of training mode

* add debug

* Do not detach when going manual

* prevent signal here

* reload training plans from disk

* remove check previous attached

* log

* Use animal current protocol here

* force re-attach

* clean

* Enh: only reload from disk when actually needed

* add on closed

* Use on_any_event with *.json

* Fix test training plan test case + clean

* comment + add test_can_copy_deepcopy_training_plan

* Delay reload plans if acquisition started

* continue test case + add reraise training-plan load error on config load/app start

* Fix: reuse reload from json

* Enh/Fix: retain raw dicts from training protocols json files, and use TrainingPlan.from_dict to spawn new plan instance

+ adapt test

* remove now unneeded test_can_copy_deepcopy_training_plan

* Adapt for dev-1 training branch

* Use slighty greater fake analysis process duration, to better see the different transitions

* enh: patch inference._intersession_process_execute when simulate_feed_intersession executed.

+++ auto-trainer-inference/src/autotrainer/inference/analysis/__init__.py
@@ -5,9 +5,9 @@ from dataclasses import dataclass
 @dataclass
 class IntersessionResponse:
     # NB: all 3 x/y/z are relative values here:
-    pellet_x: int = 0
-    pellet_y: int = 0
-    pellet_z: int = 0
+    pellet_x: float = 0
+    pellet_y: float = 0
+    pellet_z: float = 0
     food_consumed: int = 0
     successful_reaches: int = 0
     pellets_presented: int = 0
